Complete migration to iced 0.14 and localize all settings dialog strings
- Update iced::application signature and builder calls in main.rs
- Use window::latest() instead of window::get_latest() in app.rs
- Migrate keyboard subscription to keyboard::listen().filter_map in keyboard.rs
- Fix prefetch and video_trigger subscriptions to use Subscription::run with named stream functions to avoid Fn coercion/lifetime issues
- Convert INDENT_WIDTH to f32 to implement Into<Pixels> in folder_tree.rs
- Replace horizontal_space helper with Space::new().width(Length::Fill)
- Wire up all checkbox and label strings in settings_dialog.rs with l10n.tr calls

+++ crates/media-sort-gui/src/view/folder_tree.rs
@@ -6,7 +6,7 @@ use media_sort_core::models::FolderNode;
 use crate::message::Message;
 use crate::widgets::folder_icon;
 
-const INDENT_WIDTH: u16 = 20;
+const INDENT_WIDTH: f32 = 20.0;
 
 pub fn folder_tree_view<'a>(tree: &'a [FolderNode], selected_folder: Option<&'a std::path::Path>) -> Element<'a, Message> {
     column(
